feat: add docker network and volumes
