+++ database/Dockerfile
@@ -1,0 +1,2 @@
+# FROM postgres:15-alpine
+FROM postgres:16.2-alpine
