--- src/album/album.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AlbumService } from './album.service';
-
-describe('AlbumService', () => {
-  let service: AlbumService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [AlbumService],
-    }).compile();
-
-    service = module.get<AlbumService>(AlbumService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
