+++ src/album/dto/createAlbum.dto.ts
@@ -1,5 +1,5 @@
 import { IsInt, IsString, ValidateIf } from 'class-validator';
-import { ICreateAlbumDto } from 'src/types';
+import { ICreateAlbumDto } from '../helpers/types';
 
 export class CreateAlbumDto implements ICreateAlbumDto {
   @IsString()
