+++ src/album/dto/updateAlbum.dto.ts
@@ -1,5 +1,5 @@
 import { IsInt, IsString, ValidateIf } from 'class-validator';
-import { ICreateAlbumDto } from 'src/types';
+import { ICreateAlbumDto } from '../helpers/types';
 
 export class UpdateAlbumDto implements ICreateAlbumDto {
   @IsString()
