+++ src/album/helpers/types.ts
@@ -1,0 +1,5 @@
+export interface ICreateAlbumDto {
+  name: string;
+  year: number;
+  artistId: string | null; // refers to Artist
+}
