--- src/artist/artist.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ArtistController } from './artist.controller';
-
-describe('ArtistController', () => {
-  let controller: ArtistController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ArtistController],
-    }).compile();
-
-    controller = module.get<ArtistController>(ArtistController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
