+++ src/artist/dto/createArtist.dto.ts
@@ -1,5 +1,5 @@
 import { IsBoolean, IsString } from 'class-validator';
-import { ICreateArtistDto } from 'src/types';
+import { ICreateArtistDto } from '../helpers/types';
 
 export class CreateArtistDto implements ICreateArtistDto {
   @IsString()
