+++ src/artist/helpers/types.ts
@@ -1,0 +1,4 @@
+export interface ICreateArtistDto {
+  name: string;
+  grammy: boolean;
+}
