+++ src/track/dto/createTrack.dto.ts
@@ -1,5 +1,5 @@
 import { IsInt, IsString, ValidateIf } from 'class-validator';
-import { ICreateTrackDto } from 'src/types';
+import { ICreateTrackDto } from '../helpers/types';
 
 export class CreateTrackDto implements ICreateTrackDto {
   @IsString()
