+++ src/track/helpers/types.ts
@@ -1,0 +1,6 @@
+export interface ICreateTrackDto {
+  name: string;
+  artistId: string | null; // refers to Artist
+  albumId: string | null; // refers to Album
+  duration: number; // integer number
+}
