--- src/types/artist.ts
@@ -1,17 +1,0 @@
-export interface ICreateArtistDto {
-  name: string;
-  grammy: boolean;
-}
-
-export interface IArtist extends ICreateArtistDto {
-  id: string; // uuid v4
-}
-
-export enum ChangeArtistError {
-  'NOT_FOUND' = 'not found',
-}
-
-export interface IChangeArtistResult {
-  data: IArtist | null;
-  error: ChangeArtistError | null;
-}
