--- src/types/favorite.ts
@@ -1,15 +1,0 @@
-import { IAlbum } from './album';
-import { IArtist } from './artist';
-import { ITrack } from './track';
-
-export interface IFavorites {
-  artists: string[]; // favorite artists ids
-  albums: string[]; // favorite albums ids
-  tracks: string[]; // favorite tracks ids
-}
-
-export interface IFavoritesResponse {
-  artists: IArtist[];
-  albums: IAlbum[];
-  tracks: ITrack[];
-}
