--- src/types/index.ts
@@ -1,5 +1,0 @@
-export * from './user';
-export * from './artist';
-export * from './track';
-export * from './album';
-export * from './favorite';
