+++ src/user/dto/updatePassword.dto.ts
@@ -1,5 +1,5 @@
 import { IsString } from 'class-validator';
-import { IUpdatePasswordDto } from 'src/types';
+import { IUpdatePasswordDto } from '../helpers/types';
 
 export class UpdatePasswordDto implements IUpdatePasswordDto {
   @IsString()
