+++ src/user/helpers/types.ts
@@ -1,0 +1,9 @@
+export interface ICreateUserDto {
+  login: string;
+  password: string;
+}
+
+export interface IUpdatePasswordDto {
+  oldPassword: string; // previous password
+  newPassword: string; // new password
+}
